added docstrings to de-/re-hydration strategies for FormKeysValues metadata, also small qol changes
